OpenCL training
Real neural networks train on GPUs. At some point we will probably start
working on networks that cannot instantly train on CPU, my computer has
a GPU that can heavily speed this up

Pull in OpenCL as our GPU compute dependency, we briefly considered
Vulkan and OpenGL, but OpenCl felt like the best fit. A lot of the
Vulkan complexity looks like it's coming from tuning GPU pipelines. We
aren't doing any of that. This could be done with OpenGL, but OpenGL's
API is littered with assumptions made when GPUs were fixed function
pipelines for rendering. OpenCL feels like it fits nicely as an API that
gives us arbitrary GPU compute with control of how that execution is
handled

Skipping over many details as this is an initial implementation, but at
a high level...

* cl.zig is our lowest level abstraction. This basically maps 1-1 to
  OpenCl apis
* math.zig holds our new abstraction of a Tensor. This is an ND matrix
  that can be backed by many different buffer types. This is useful as
  we can have different tensor types for traced/untraced execution that
  we can expect to use similarly
* We stack an abstraction on top as math.Executor. This is our math
  execution engine that wraps the low level cl.Executor. This basically
  holds all our kernels and maps arguments into/off the GPU
* One more layer comes on top, TracingExecutor. This wraps any
  math.Executor APIs that we need for training. His job is to build a
  graph of math operations. We use a different Tensor type here that is
  backed by node ids instead of the OpenCL buffer
* We now do training with matrices instead of individually tracking
  floats. This means less graph tracking for more numbers, yay!