Bounding boxes
The current idea for our final barcode scanner is to isolate a barcode
from a larger image, blow it up, and use the high res image of just the
barcode to do parsing. The first step of this process is to figure out
where the barcode is, and a bounding box is a nice way to do so

Do an initial attempt at bounding box prediction. Note that this is
known to make a not so great predictor. Future changes will make
experimentation and tuning easier, so just commit a baseline for now

Supporting changes
* Predict sqrt w/h to allow larger error in larger boxes, yolo v1
  inspired
* Backgrounds temporarily disabled as we experiment to make problem
  easier
* All references to orientations removed, as orientation is exported via
  bbox
* Wavy bars temporarily disabled as it is hard to know how the box
  should be drawn when the ripples push the edge of the bar out of it's
  initial w/h box. Something more intelligent to adjust w/h based on max
  ripple may be needed
* relu upgraded to leaky relu. We observed lack of gradient propagation
  during training, adding a leak (like in yolo v1) seemed to nicely
  resolve this
* Even contribution of x/y/w/h/rx/ry resulted in poor training. Early
  experimentation indicated that 0 rotation resulted in reasonable w/h
  predictions. Implement elemMul and use it on loss, reducing
  contribution of rx/ry as well as height as incorrect height should be
  a much smaller problem than incorrect w/cx/cy
* Tired of copying around FIXMES that we aren't doing in
  TrainingExecutor, remove the ones we are copying and aren't doing