Runtime layer definition
We are at the point now where we are unsure of what a good model
definition is for our bbox predictor. Experimentation via config is much
more flexible. In the future we will be able to save
configs/logs/checkpoints for multiple networks and compare. This also
unlocks training multiple networks in a row with the same compiled code.
Tracking differences via config should also be significantly easier

Config is oddly stored as a python file. JSON is nice in that it's in
the standard lib, and everything on earth supports it. It is not nice in
that it is a shit format. We store our config as python so that we can
do math in it, add comments, etc. This can be passed into our program as
--config <(./default_config.py) to convert into a temporary file
descriptor

Config format is split between things that are part of our main loop,
and things that are owned by nn.zig with the intent of eventually having
a generic nn framework.

Supporting changes:
* Drop batch size from reshape layer, config needing to know batch size
  to define model seems kinda wrong
  * Layer execution now needs to merge batch size with the saved config
* Reshape needs to copy shape now, because config shape is temporary.
  Previously reshapes always referenced compile time constants
* Weight initialization used to take anytypes of the initializer.
  Initializers are no longer known at compile time so we need an
  nn.Initializer
  * This makes an Initializers type convenient. Create one of each and
    resolve from config later
* Some global constants (initial lr, batch size, img size) are now part
  of config. Parse config in main and send to training thread. Pass
  these parameters around where they need to go
* OpenClLayerGen finally removed
  * Model construction is now owned by nn.zig. The way the code ended
    up, we don't need this struct to hold parameters for us. Make
    functions for each layer that just export nn.Layer(Executor)
    directly
* Tracing executor max operations * 10. This is a relatively cheap
  allocation anyways, and we were hitting our limits with some network
  definitions