Collect and visualize training statistics
We are past the point now where we can eyeball if a change is good or
not, we need information to make further decisions about network
architecture etc.

During training, output a log file for consumption by a new data
visualizer. This data is basically a tagged csv. It can log arbitrary
info at arbitrary intervals

Log the loss every N iters. Add a validation set and check performance
on that every Y iters

The validation stuff is just inlined in the training loop for now. It's
far from optimal, but we have run many experiments this way. In the
interest of getting those experiments merged, we accept some sub-optimal
code in this commit ;)