Heal orientations experiment
We have noticed that disabling rotation loss results in much better
predictions for the rest of the network. Looking at some of our stats,
it seems like we are really struggling to predict the Y vector of our
rotation. This makes sense as there is still near symmetry if all our
vectors are on the right side

Add option to heal labels to match predictions. Flip our label if the
prediction seems to be going in the opposite direction.

A little shuffling needed to be done to ensure the notifier was
notifying post-healed bbox labels