Add bar prediction training target
It's time to start running some experiments for the second stage of the
barcode parser. Export the barcode as a buffer that we can use as a
training target. Update loss/validation to not fall over in this mode

For now the stats are pretty weak, this is just a quick and dirty
initial attempt to get us moving