Kill app when training thread dies
If initialization fails, our gui just sits there drooling and doing
nothing. Add a shutdown message that can be sent from the training
thread to the gui thread, and send it on thread exit

On thread exit, it's easy for the GUI to reference data that was owned
by the training thread. Copy layer names out to avoid this dependency