Alternative confidence metric
IOU as confidence does not seem to nicely bisect good and bad boxes.
Maybe rotation error will behave better

A lot of renames from iou -> confidence. Bool params are promoted to
enum/int params. For now opencl just uses ints that are hardcoded in the
barcodgen enum, but we could mirror the enum if we felt that made more
sense