Bbox validation improvements
After running some experiments, sometimes we want to look at data that
we didn't collect while training. We need a way to run validation and
visualize it on the final checkpoints of many different training runs

Split out bbox validation stats, as that's what we need at this moment.
These have been re-written to show the things we care about. Create a
new executable that runs the stats and creates a csv of the outputs. For
now these csvs can be imported into a spreadsheet for comparison, but a
long term goal is to merge this with the training_output_vis.

This has only been done for box validation, but longer term we probably
also want to give a similar treatment to the bars stats. Commit as is
for now as it's a step in the right direction, even if not fully
complete