Optionally disable bbox loss when out of frame
Enable an experiment to see if disabling box loss when out of frame
improves performance. When we increase our randomization to show more
out of frame boxes, our performance on real data gets worse. Can we keep
it the same by just not adjusting our bbox loss in this case? Only one
way to find out

Committing a relatively hacky solution that just does this in label
healing by setting predictions to expected to force 0 loss. We might
want a better solution if this is deemed to be a good thing to do