Update default parse network
With this network we achieve better accuracy with faster execution and
smaller weight sizes. We have an example of getting 100% correct in
scanner.zig after 4 hours of training. The network has not peaked yet,
but training for longer at this stage of development is a waste of time

Our weight sizes are still ~10M, which is very large. It seems like we
can cut this down more, but this is already a 15x reduction almost, so
take the win we have

Changes made...
* Re-jig data generation to pick any arbitrary orientation in the center
  of an image. Size the image such that we are ~ equivalent to a
  1024x1024 image with default_bbox params
* Enable extraction. Randomization parameters are fairly arbitrary, just
  pick stuff that looks about right
* Extracted size picked such that
  * Initial input looks reasonable to my human eyes
  * Pooling brings us to a 95x6 image before our fully connected layer
    * Assumption here is that we still want some vertical data for
      global reasoning, but most of it can be pre-processed by earlier
      convolutional layers such that we don't need a ton
* Pool vertically more, and more often